+++ CHANGES.md
@@ -62,3 +62,4 @@
 0.1.15
 * add relay restrictions - required authenticaiton and list of users/domains
 * fix opam lint errors
+* add dedup to calculate duplicate attachments in mbox